scripts/requirements.txt
# Core dependencies for RLS troubleshooting RAG system
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
ujson>=5.8.0
numpy>=1.24.0
anthropic>=0.18.0
rank-bm25>=0.2.2
python-dotenv>=1.0.0

